I2S logging compile warning

+++ src/AudioI2S/I2SESP32V1.h
@@ -202,7 +202,7 @@ class I2SDriverESP32V1 {
       int size = cfg.buffer_size * cfg.buffer_count;
       int frame_size = cfg.bits_per_sample * cfg.channels / 8;
       if (size > 0) result.dma_frame_num = size / frame_size;
-      LOGI("dma_frame_num: %d", result.dma_frame_num);
+      LOGI("dma_frame_num: %d", (int)result.dma_frame_num);
       result.auto_clear = cfg.auto_clear;
       return result;
     }
